Fix model training, execution, and feature collinearity issues
- trading_model: Pass DataFrame instead of .values to preserve feature
  names in sklearn pipelines; drop 11 collinear features (61→50) based
  on VIF/correlation analysis
- live: Improve dust sell handling with warning instead of silent skip;
  add quantity rounding to avoid Binance precision errors
- position_sizing: Thread min_confidence parameter through to sizer
- sentiment_advanced: Fix .astype(float) on regime numeric column

+++ src/features/sentiment_advanced.py
@@ -409,7 +409,7 @@ class AdvancedSentimentAnalyzer:
             'bullish': 1,
             'very_bullish': 2
         }
-        df['sentiment_regime_numeric'] = df['sentiment_regime_level'].map(regime_map)
+        df['sentiment_regime_numeric'] = df['sentiment_regime_level'].map(regime_map).astype(float)
 
         # Momentum-based regime
         if 'sentiment_velocity' in df.columns:
